+++ M80ImageMerger/M80ImageMerger/Classes/Algorithm/M80ImageMergeInfo.m
@@ -91,11 +91,11 @@
 
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"%@ 1st height %lf offset %zd 2nd height %lf offset %zd length %zd"
+    return [NSString stringWithFormat:@"%@ 1st height %lf offset %d 2nd height %lf offset %d length %d"
             ,_type == M80FingerprintTypeCRC ? @"crc" : @"hist"
-            ,_firstImage.size.height,_firstOffset,
-            _secondImage.size.height,_secondOffset,
-            _length];
+            ,_firstImage.size.height,(int)_firstOffset,
+            _secondImage.size.height,(int)_secondOffset,
+            (int)_length];
 }
